+++ MiamIOSFramework/MiamIOSFramework/ui/BasketPreview/BasketPreviewRow.swift
@@ -19,7 +19,7 @@ struct BasketPreviewRow: View {
     private let productPrice: String
     private var viewModel : BasketPreviewVM
     private let previewLine: BasketPreviewLine
-    private let showChangeProductLink : Bool
+    private let itemsCount : Int
     
     private var removeProductAction: () -> Void = {}
     private var replaceProductAction: () -> Void = {}
@@ -46,8 +46,7 @@ struct BasketPreviewRow: View {
         }
         self.removeProductAction = removeProductAction
         self.replaceProductAction = replaceProductAction
-        self.showChangeProductLink = !((previewLine.record as! BasketEntry).relationships?.items?.data.isEmpty ?? true ||
-                                       (previewLine.record as! BasketEntry).relationships!.items!.data.count == 1 )
+        self.itemsCount = (previewLine.record as! BasketEntry).relationships?.items?.data.count ?? 0
     }
     
     func onQuantityChanged(value: Int){
@@ -62,7 +61,7 @@ struct BasketPreviewRow: View {
         
          
         if (Template.sharedInstance.basketPreviewRowTemplate != nil) {
-            Template.sharedInstance.basketPreviewRowTemplate!(productName, productPictureURL, productBrandName, productName, productPrice, Int(previewLine.count),showChangeProductLink, removeProductAction, replaceProductAction, onQuantityChanged)
+            Template.sharedInstance.basketPreviewRowTemplate!(productName, productPictureURL, productBrandName, productName, productPrice, Int(previewLine.count),itemsCount, removeProductAction, replaceProductAction, onQuantityChanged)
         } else {
             VStack(alignment: .leading) {
                 HStack(alignment: .top) {
@@ -125,7 +124,7 @@ struct BasketPreviewRow: View {
                 //Ingredeient View
                 HStack {
                     HStack {
-                        if(showChangeProductLink){
+                        if(itemsCount > 1){
                             Button(action: {
                                 replaceProductAction()
                             }) {