+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/catalog/CatalogPage.kt
@@ -1,6 +1,7 @@
 package com.miam.kmm_miam_sdk.android.ui.components.catalog
 
 import android.content.Context
+import android.util.AttributeSet
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -22,11 +23,13 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
+import androidx.compose.ui.platform.AbstractComposeView
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import com.miam.kmmMiamCore.component.recipeListPage.RecipeListPageContract
 import com.miam.kmmMiamCore.component.recipeListPage.RecipeListPageViewModel
+import com.miam.kmmMiamCore.handler.LogHandler
 import com.miam.kmmMiamCore.miam_core.model.Recipe
 import com.miam.kmm_miam_sdk.android.theme.Colors
 import com.miam.kmm_miam_sdk.android.theme.Colors.primary
@@ -45,148 +48,140 @@ import com.miam.kmm_miam_sdk.android.ui.components.favoritePage.FavoritePageStyl
 import com.miam.kmm_miam_sdk.android.ui.components.recipeCard.RecipeView
 import com.miam.kmm_miam_sdk.android.ui.components.states.ManagementResourceState
 
-@Composable
-fun CatalogPage(
-    recipePageVM: RecipeListPageViewModel,
+open class CatalogPage @JvmOverloads constructor(
     context: Context,
-    columns: Int,
-    verticalSpacing: Int,
-    horizontalSpacing: Int,
-    returnToCategoriesPage: () -> Unit
-) {
+    attrs: AttributeSet? = null,
+    defStyleAttr: Int = 0
+): AbstractComposeView(context, attrs, defStyleAttr) {
 
-    val state = recipePageVM.uiState.collectAsState()
+    private val recipePageVM = RecipeListPageViewModel()
+    private var catalogPageTitle = ""
+    private var catalogPageColumns = 1
+    private var catalogPageVerticalSpacing = 12
+    private var catalogPageHorizontalSpacing = 12
+    private var catalogPageBack: () -> Unit = { LogHandler.error("[Miam][Error] you must bind back function") }
 
-    ManagementResourceState(
-        resourceState = state.value.recipes,
-        successView = { recipes ->
-            requireNotNull(recipes)
-            CatalogSuccessPage(
-                recipePageVM,
-                recipes,
-                columns,
-                verticalSpacing,
-                horizontalSpacing,
-                context
-            )
-        },
-        emptyView = {
-            CatalogEmptyPage(
-                recipePageVM.currentState.title,
-                recipePageVM.currentState.filter.contains("filter[liked]=true&"),
-                returnToCategoriesPage
-            )
-        },
-        loadingView = {
-            if (Template.CatalogResultPageLoadingTemplate != null) {
-                Template.CatalogResultPageLoadingTemplate?.let {
-                    it()
-                }
-            } else {
-                CatalogLoadingPage()
-            }
-        },
-        onTryAgain = { /**TODO*/ },
-        onCheckAgain = { /**TODO*/ },
-    )
-}
+    fun bind(
+        title: String,
+        back: () -> Unit,
+        columns: Int? = null,
+        verticalSpacing: Int? = null,
+        horizontalSpacing: Int? = null
+    ) {
+        catalogPageTitle = title
+        catalogPageBack = back
+        columns?.let { catalogPageColumns = it }
+        verticalSpacing?.let { catalogPageVerticalSpacing = it }
+        horizontalSpacing?.let { catalogPageHorizontalSpacing = it }
+        recipePageVM.setEvent(RecipeListPageContract.Event.InitPage(title))
+    }
 
-@Composable
-private fun CatalogSuccessPage(
-    recipePageVM: RecipeListPageViewModel,
-    recipes: List<Recipe>,
-    columns: Int,
-    verticalSpacing: Int,
-    horizontalSpacing: Int,
-    context: Context
-) {
-    Box(modifier = Modifier.fillMaxSize()) {
-        LazyVerticalGrid(
-            modifier = FavoritePageStyle.favoriteMainContainer,
-            columns = GridCells.Fixed(columns),
-            verticalArrangement = Arrangement.spacedBy(verticalSpacing.dp, Alignment.Top),
-            horizontalArrangement = Arrangement.spacedBy(horizontalSpacing.dp, Alignment.Start)
-        ) {
-            item(span = { GridItemSpan(columns) }) {
-                if (Template.CatalogPageTitleTemplate != null) {
-                    Template.CatalogPageTitleTemplate?.let {
-                        it(recipePageVM.currentState.title)
-                    }
-                } else {
-                    Row {
-                        Text(
-                            text = recipePageVM.currentState.title,
-                            color = Colors.black,
-                            style = Typography.subtitleBold
-                        )
-                    }
-                }
-            }
-            itemsIndexed(recipes) { index, item ->
-                val recipe = RecipeView(context = context)
-                recipe.bind(recipe = item)
-                recipe.isNotInShelf()
-                recipe.Content()
-                if (index == recipes.lastIndex) {
+    @Composable
+    override fun Content() {
+        val state = recipePageVM.uiState.collectAsState()
+
+        ManagementResourceState(
+            resourceState = state.value.recipes,
+            successView = { recipes ->
+                requireNotNull(recipes)
+                CatalogSuccessPage(
+                    context,
+                    catalogPageTitle,
+                    recipes,
+                    catalogPageColumns,
+                    catalogPageVerticalSpacing,
+                    catalogPageHorizontalSpacing,
+                    recipePageVM.currentState.isFetchingNewPage,
+                ) {
                     recipePageVM.setEvent(RecipeListPageContract.Event.LoadPage)
                 }
-            }
-            item(span = { GridItemSpan(columns) }) {
-                if (recipePageVM.currentState.isFetchingNewPage) {
-                    if (Template.CatalogResultPageLazyLoaderTemplate != null) {
-                        Template.CatalogResultPageLazyLoaderTemplate?.let {
-                            it()
-                        }
-                    } else {
-                        Row(
-                            modifier = FavoritePageStyle.loadMoreContainer,
-                            Arrangement.Center
-                        ) {
-                            CircularProgressIndicator(
-                                color = FavoritePageColor.loaderColor,
-                                modifier = FavoritePageStyle.loadMoreModifier
-                            )
-                        }
+            },
+            emptyView = {
+                CatalogEmptyPage(
+                    recipePageVM.currentState.title,
+                    recipePageVM.currentState.filter.contains("filter[liked]=true&"),
+                    catalogPageBack
+                )
+            },
+            loadingView = {
+                if (Template.CatalogResultPageLoadingTemplate != null) {
+                    Template.CatalogResultPageLoadingTemplate?.let { it() }
+                } else {
+                    CatalogLoadingPage()
+                }
+            },
+            onTryAgain = { /**TODO*/ },
+            onCheckAgain = { /**TODO*/ },
+        )
+    }
+
+    @Composable
+    private fun CatalogSuccessPage(
+        context: Context,
+        title: String,
+        recipes: List<Recipe>,
+        columns: Int,
+        verticalSpacing: Int,
+        horizontalSpacing: Int,
+        isFetching: Boolean,
+        loadPage: () -> Unit
+    ) {
+        Box(modifier = Modifier.fillMaxSize()) {
+            LazyVerticalGrid(
+                modifier = FavoritePageStyle.favoriteMainContainer,
+                columns = GridCells.Fixed(columns),
+                verticalArrangement = Arrangement.spacedBy(verticalSpacing.dp, Alignment.Top),
+                horizontalArrangement = Arrangement.spacedBy(horizontalSpacing.dp, Alignment.Start)
+            ) {
+                item(span = { GridItemSpan(columns) }) {
+                    HeaderTitle(title = title)
+                }
+                itemsIndexed(recipes) { index, item ->
+                    val recipe = RecipeView(context = context)
+                    recipe.bind(recipe = item)
+                    recipe.isNotInShelf()
+                    recipe.Content()
+                    if (index == recipes.lastIndex) {
+                        loadPage()
                     }
+                }
+                item(span = { GridItemSpan(columns) }) { FooterLoader(isFetching) }
+            }
+        }
+    }
+
+    @Composable
+    private fun HeaderTitle(title: String) {
+        val templateToUse = specificTemplate() ?: Template.CatalogPageTitleTemplate
+        if (templateToUse != null) {
+            templateToUse(title)
+        } else {
+            Row { Text(text = title, color = Colors.black, style = Typography.subtitleBold) }
+        }
+    }
+
+    protected open fun specificTemplate(): @Composable() ((String) -> Unit)? {
+        return null
+    }
+
+    @Composable
+    private fun FooterLoader(isFetching: Boolean) {
+        if (isFetching) {
+            if (Template.CatalogResultPageLazyLoaderTemplate != null) {
+                Template.CatalogResultPageLazyLoaderTemplate?.let { it() }
+            } else {
+                Row(modifier = FavoritePageStyle.loadMoreContainer, Arrangement.Center) {
+                    CircularProgressIndicator(color = FavoritePageColor.loaderColor, modifier = FavoritePageStyle.loadMoreModifier)
                 }
             }
         }
     }
-}
 
-@Composable
-private fun CatalogLoadingPage() {
-    Column(
-        Modifier.fillMaxSize(),
-        horizontalAlignment = Alignment.CenterHorizontally,
-        verticalArrangement = Arrangement.Center
-    ) {
-        Text(recipesListLoading, style = Typography.subtitleBold, modifier = Modifier.padding(8.dp))
-        CircularProgressIndicator(color = Colors.primary)
-    }
-}
-
-@Composable
-fun CatalogEmptyPage(
-    title: String,
-    isFavorite: Boolean,
-    returnToCategoriesPage: () -> Unit
-) {
-    if (isFavorite) {
-        if (Template.CatalogFavoritEmptyTemplate != null) {
-            Template.CatalogFavoritEmptyTemplate?.let {
-                it(returnToCategoriesPage)
-            }
-        } else {
-            FavoriteEmptyPage(returnToCategoriesPage)
-        }
-    } else {
-        if (Template.CatalogSearchResultEmptyTemplate != null) {
-            Template.CatalogSearchResultEmptyTemplate?.let {
-                it(returnToCategoriesPage, title)
-            }
-        } else {
-            SearchResultEmptyPage(pageTitle = title)
+    @Composable
+    private fun CatalogLoadingPage() {
+        Column(Modifier.fillMaxSize(), horizontalAlignment = Alignment.CenterHorizontally, verticalArrangement = Arrangement.Center) {
+            Text(recipesListLoading, style = Typography.subtitleBold, modifier = Modifier.padding(8.dp))
+            CircularProgressIndicator(color = Colors.primary)
         }
     }
 }
@@ -204,13 +199,8 @@ fun FavoriteEmptyPage(returnToCategoriesPage: () -> Unit) {
                 .fillMaxWidth(),
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            Image(
-                painter = painterResource(CatalogImage.empty),
-                contentDescription = null,
-                Modifier
-                    .padding(vertical = 16.dp)
-            )
-            Clickable(onClick = { returnToCategoriesPage() }) {
+            Image(painter = painterResource(CatalogImage.empty), contentDescription = null, Modifier.padding(vertical = 16.dp))
+            Clickable(onClick = returnToCategoriesPage) {
                 Column {
                     Text(
                         text = noFavoriteYet,
@@ -221,10 +211,7 @@ fun FavoriteEmptyPage(returnToCategoriesPage: () -> Unit) {
                             .padding(horizontal = 16.dp)
                             .padding(bottom = 8.dp)
                     )
-                    Row(
-                        Modifier.fillMaxWidth(),
-                        horizontalArrangement = Arrangement.Center
-                    ) {
+                    Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.Center) {
                         Box(
                             Modifier
                                 .clip(RoundedCornerShape(50))
@@ -234,10 +221,7 @@ fun FavoriteEmptyPage(returnToCategoriesPage: () -> Unit) {
                                 Text(
                                     text = backToCategories,
                                     color = primary,
-                                    modifier = Modifier.padding(
-                                        horizontal = 16.dp,
-                                        vertical = 8.dp
-                                    ),
+                                    modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp),
                                 )
                                 Image(
                                     painter = painterResource(CatalogImage.back),
@@ -266,12 +250,7 @@ fun SearchResultEmptyPage(pageTitle: String) {
                 .fillMaxWidth(),
             horizontalAlignment = Alignment.CenterHorizontally
         ) {
-            Image(
-                painter = painterResource(CatalogImage.empty),
-                contentDescription = null,
-                Modifier
-                    .padding(vertical = 16.dp)
-            )
+            Image(painter = painterResource(CatalogImage.empty), contentDescription = null, Modifier.padding(vertical = 16.dp))
             Text(
                 text = noResultFor + pageTitle,
                 color = white,
@@ -279,10 +258,30 @@ fun SearchResultEmptyPage(pageTitle: String) {
                 textAlign = TextAlign.Center,
                 modifier = Modifier.padding(horizontal = 16.dp)
             )
-            Text(
-                text = tryAnOtherSearch,
-                color = white
-            )
+            Text(text = tryAnOtherSearch, color = white)
         }
     }
 }
+
+
+@Composable
+fun CatalogEmptyPage(
+    title: String,
+    isFavorite: Boolean,
+    returnToCategoriesPage: () -> Unit
+) {
+    if (isFavorite) {
+        if (Template.CatalogFavoritEmptyTemplate != null) {
+            Template.CatalogFavoritEmptyTemplate?.let { it(returnToCategoriesPage) }
+        } else {
+            FavoriteEmptyPage(returnToCategoriesPage)
+        }
+    } else {
+        if (Template.CatalogSearchResultEmptyTemplate != null) {
+            Template.CatalogSearchResultEmptyTemplate?.let { it(returnToCategoriesPage, title) }
+        } else {
+            SearchResultEmptyPage(pageTitle = title)
+        }
+    }
+}
+