+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/catalogHeader/CatalogHeader.kt
@@ -73,10 +73,8 @@ fun CatalogHeader(
         headerState = HeaderState.PREFERENCES
     }
 
-    // TODO Refact with filter service
-    val filter = CatalogFilter(catalogFilterVm, ::closeModal, goToRecipeList)
-    // TODO Refact with filter service
-    val search = CatalogSearch(catalogFilterVm, ::closeModal, goToRecipeList)
+    val filter = CatalogFilter(::closeModal, goToRecipeList)
+    val search = CatalogSearch(::closeModal, goToRecipeList)
 
 
     fun getActiveFilterCount(): Int {