+++ miamCore/src/commonMain/kotlin/com/miam/kmmMiamCore/component/singletonFilter/SingletonFilterViewModel.kt
@@ -19,7 +19,7 @@ open class SingletonFilterViewModel:
     private val recipeRepositoryImp: RecipeRepositoryImp by inject()
 
     private val coroutineHandler = CoroutineExceptionHandler { _, exception ->
-        LogHandler.error("Miam error in catalog view $exception")
+        LogHandler.error("Miam error in singleton filter VM view $exception")
     }
 
     init {
@@ -87,21 +87,19 @@ open class SingletonFilterViewModel:
     fun setCat(catId: String) {
         setState { copy(category = catId) }
         getRecipeCount()
-        @Suppress("unused until 3.0.0")
-        setEffect { SingletonFilterContract.Effect.OnUpdate }
     }
 
     fun setFavorite() {
         setState { copy(isFavorite = true) }
         getRecipeCount()
-        @Suppress("unused until 3.0.0")
-        setEffect { SingletonFilterContract.Effect.OnUpdate }
     }
 
     fun setSearchString(searchString: String) {
         setState { copy(searchString = searchString) }
         getRecipeCount()
-        @Suppress("unused until 3.0.0")
+    }
+
+    fun applyFilter() {
         setEffect { SingletonFilterContract.Effect.OnUpdate }
     }