+++ changelog.md
@@ -1,6 +1,12 @@
 # 3.5.0
-- [FIX][IOS] Add onQuantityChanged closure to basketPreviewRowTemplate and change item link avaibility
-- [FEA][COR] Preference dependence on supplier 
+- [FIX][COR] Fix navigation when a dialog was between two pages
+- [FEA][COR] Preference dependence on supplier
+- [FEA][AND] Comfirme prefrerences now close dialog instead of redirecting to a search result
+- [FEA][AND] CatalogSearch is now a stand alone component 
+- [FIX][IOS] Add onQuantityChanged closure to basketPreviewRowTemplate and change item link avaibility 
+- [FIX][IOS] package subtitle exposed
+- [FIX][IOS] MyMealButton onTapped closure used
+
 
 # 3.4.2
 - [FIX][COR] Like initial state